Fixes for validations, especially late (post-settle) 1. Invalid payloads for an Inference will fail validation. 2. Validators with no nodes for the model will not submit a validation. 3. validations that are submitted after the settle will not get a share of the work, but will still count for inferences validated. 4. invalidations voted on and submitted after the settle will not be refunded. 5. Mock server for tests can now return an error at a validation attempt, allowing us to delay validations until after settle to test the above changes. 6. Support for direct submission of MsgValidation and MsgClaimRewards in tests 7. Support for retrieving and serializing app-config.yaml from API nodes in tests. Main purpose is to get Seed values for MsgClaimRewards. 8. Enabling some validation tests now. Some remain unstable

+++ inference-chain/x/inference/keeper/accountsettle_test.go
@@ -162,7 +162,7 @@ func TestSingleSettle(t *testing.T) {
 	result, newCoin, err := inference.GetSettleAmounts([]types.Participant{participant1}, &defaultSettleParameters)
 	require.NoError(t, err)
 	require.Equal(t, 1, len(result))
-	require.Equal(t, expectedRewardCoin, uint64(newCoin.Amount))
+	require.Equal(t, expectedRewardCoin, newCoin.Amount)
 	p1Result := result[0]
 	require.Equal(t, uint64(1000), p1Result.Settle.WorkCoins)
 	require.Equal(t, uint64(expectedRewardCoin), p1Result.Settle.RewardCoins)
@@ -183,7 +183,7 @@ func TestEvenSettle(t *testing.T) {
 	result, newCoin, err := inference.GetSettleAmounts([]types.Participant{participant1, participant2}, &defaultSettleParameters)
 	require.NoError(t, err)
 	require.Equal(t, 2, len(result))
-	require.Equal(t, expectedRewardCoin, uint64(newCoin.Amount))
+	require.Equal(t, expectedRewardCoin, newCoin.Amount)
 	p1Result := result[0]
 	require.Equal(t, uint64(1000), p1Result.Settle.WorkCoins)
 	require.Equal(t, uint64(expectedRewardCoin/2), p1Result.Settle.RewardCoins)
